projects/deliveries_api/pyproject.toml
[tool.poetry]
name = "deliveries_api"
version = "0.1.0"
description = ""
authors = ["Bernardo Gomes <bgomesdeabreu@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.7"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
django-model-utils = "^4.1.1"
psycopg2 = "^2.9.1"
prettyconf = "^2.2.1"
kafka-python = "^2.0.2"
django-simple-history = "^3.0.0"
django-extensions = "^3.1.3"
drf-yasg = "^1.20.0"
gunicorn = "^20.1.0"
sentry-sdk = "^1.4.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
